Orbital: Handle incorrectly encoded response
When an order contains non-standard characters (normally in the
customer's name and is normally accented characters) and is submitted
to Orbital, they will send back an XML response flagged as UTF-8 but is
actually ISO-8859-1. This causes parsers like REXML to raise an
exception while parsing what it thinks is UTF-8.